Abstract reconciliaiton functions
